BOTS/08.PDF_Parser/requirements.txt
ollama==0.6.1
python-dotenv==1.0.0
pydantic==2.12.5
PyMuPDF==1.26.6
requests==2.32.5
numpy==1.26.4
Pillow==10.0.0
fastapi==0.121.2
uvicorn==0.38.0
websockets==15.0.1
python-multipart==0.0.20
streamlit==1.41.0